# Import speed improvements