This includes a few changes to improve the import speed of maps, models and entities, although it's way more noticeable in maps, this PR being focused on that.

First is a tiny weeny thing to remove a `str.lower()` in bin_ops which for map imports ran millions of times and used up to a minute according to profiling. Second is a selection tracking function to avoid calling `bpy.ops.select_all()` thousands of times causing Blender to update the scene unnecessarily. Third, are some changes in `importEmpties` to avoid using `bpy.ops` causing unnecessary updates again; this apparently fixed an issue with empties in map entities having wrong positions. 

There are further optimizations to be made in `importArmature` but I noped straight out of that code, honestly.
Also, the selection tracking was not implemented in the effects importer, that's still using UI operators.

These were measured on a I7-4790k with Blender 2.92

| Map  | Old Import Time | New Import Time |
| ------------- | ------------- | ------------- |
| Menu Hanamura  | 00:39  | 00:24  |
| Menu King's Row  | 01:36 |  00:52 |
| Menu Blizz World (W)  | 01:50 |  00:56 |
| Black Forest  | 04:00 |  02:13 |
| Practice Range  | 07:35 | 04:26  |
| Ilios Well  | 10:41 |  05:15 |